refactoring: moved WithTokenGetRequestProcessor
WithTokenGetRequestProcessor was moved from
web/controller package to web/security package,
since this class is highly related to application
security level.

Since WithTokenGetRequestProcessor was scanned by WebMvcConfig,
after moving to another dir this bean should be defined
manually or scanned in SecurityMock config, so it was added.

Note that real implementation is used, as it was before this change,
it probably should be replaced by mock in future.